scam/requirements.txt
streamlit
joblib
python-docx
PyMuPDF
scikit-learn
